Move core functionality into runtime subpkg
