+++ package.json
@@ -7,9 +7,11 @@
     "build:core": "microbundle --raw --cwd packages/core && pnpm postbuild:core",
     "build:preact": "microbundle --raw --cwd packages/preact && pnpm postbuild:preact",
     "build:react": "microbundle --raw --cwd packages/react && pnpm postbuild:react",
+    "build:react-runtime": "microbundle --raw --cwd packages/react/runtime && pnpm postbuild:react-runtime",
     "postbuild:core": "cd packages/core/dist && mv -f index.d.ts signals-core.d.ts",
     "postbuild:preact": "cd packages/preact/dist && mv -f preact/src/index.d.ts signals.d.ts && rm -dr preact",
     "postbuild:react": "cd packages/react/dist && mv -f react/src/index.d.ts signals.d.ts && rm -dr react",
+    "postbuild:react-runtime": "cd packages/react/runtime/dist && mv -f react/runtime/src/*.d.ts . && rm -dr react",
     "postbuild": "node ./scripts/node-13-exports.js",
     "lint": "pnpm lint:eslint && pnpm lint:tsc",
     "lint:eslint": "eslint 'packages/**/*.{ts,tsx,js,jsx}'",
