+++ packages/react/runtime/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@preact/signals-react-runtime",
+  "description": "Sub package for @preact/signals-react that contains the pure runtime functions",
+  "private": true,
+  "amdName": "reactSignalsRuntime",
+  "main": "dist/runtime.js",
+  "module": "dist/runtime.module.js",
+  "unpkg": "dist/runtime.min.js",
+  "types": "dist/index.d.ts",
+  "source": "src/index.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "browser": "./dist/runtime.module.js",
+      "umd": "./dist/runtime.umd.js",
+      "import": "./dist/runtime.mjs",
+      "require": "./dist/runtime.js"
+    }
+  },
+  "dependencies": {
+    "@preact/signals-core": "workspace:^1.3.0",
+    "use-sync-external-store": "^1.2.0"
+  },
+  "peerDependencies": {
+    "react": "^16.14.0 || 17.x || 18.x"
+  }
+}
